+++ example/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+ - Add PreferencePage for setting theme, to demonstrate that the child page will react to theme changes.
 ### Changed
  - Update dependencies.
  - Null safety.